Spread base classes into their modules.
