+++ docs/api.rst
@@ -209,5 +209,4 @@ API
    nimare.base.base
    nimare.base.annotate
    nimare.base.decode
-   nimare.base.meta
    nimare.base.misc
