+++ nimare/annotate/topic/boltzmann.py
@@ -1,7 +1,7 @@
 """
 Topic modeling with deep Boltzmann machines.
 """
-from ...base import AnnotationModel
+from ..base import AnnotationModel
 from ...due import due
 from ... import references
