+++ nimare/annotate/vector/word2brain.py
@@ -1,7 +1,7 @@
 """
 GloVe model-based annotation.
 """
-from ...base import AnnotationModel
+from ..base import AnnotationModel
 from ...due import due
 from ... import references
