+++ nimare/base/__init__.py
@@ -3,12 +3,10 @@ Base classes used throughout NiMARE
 """
 
 from .base import MetaResult
-from .decode import Decoder
-from .annotate import AnnotationModel
-from .misc import Parcellator
+from .estimators import Estimator, Transformer
 
-__all__ = ['Decoder',
-           'MetaResult',
-           'AnnotationModel',
-           'Parcellator'
-           ]
+__all__ = [
+    'Estimator',
+    'Transformer',
+    'MetaResult'
+]
