+++ nimare/parcellate/cbp.py
@@ -8,7 +8,7 @@ from scipy.spatial.distance import cdist
 import scipy.ndimage.measurements as meas
 from nilearn.masking import apply_mask, unmask
 
-from ..base import Parcellator
+from .base import Parcellator
 from ..meta.cbma.ale import SCALE
 from ..due import due
 from .. import references
